requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pymysql==1.1.1
python-jose==3.3.0
python-dotenv==1.0.1

